perf+security+quality: comprehensive 49-point optimization
Security:
- Auth body size limit: usize::MAX -> 1MB on auth endpoints
- Token entropy check: require mixed alphanumeric chars
- Fix batch_delete_videos storage quota decrement bug

Backend Performance:
- Parallelize comment count+fetch with tokio::try_join!
- Deduplicate Cloudflare IP ranges (shared module)
- Const IP address parsing (no runtime .parse())
- Optimize sanitize_for_log with Cow<str> zero-alloc fast path

Frontend Performance:
- Remove redundant startPlaybackSession call (2 -> 1 per video load)
- Clean up console.log in production code
- Fix 'as any' type casts with proper types
- Remove duplicate mousedown listener (pointerdown only)
- Delete dead code (SearchIntegration.tsx)

i18n:
- Category names now translated (zh-CN/en-US)
- VideoCard views text uses i18n keys
- Player hardcoded strings removed

UX/Accessibility:
- VideoCard shows '新' badge instead of '0 次观看'

Code Quality:
- Deduplicate is_valid_email function
- Standardize cloudflare_ips shared module

API Design:
- Paginate /videos/favorites endpoint
- Paginate /playback/history endpoint
- Update OpenAPI spec for new pagination params

Tests:
- Add playlist integration tests (6 test cases)
- Fix layout test for pointerdown event
- 30 test files, 574 tests all passing

+++ backend/tests/service_misc_tests.rs
@@ -339,10 +339,10 @@ async fn search_nonempty_query_reaches_db() {
 //     失败仅记日志、不入队副作用，正好验证"内存队列不依赖 DB"的契约。
 // ══════════════════════════════════════════════════════════════════════
 
-fn make_task(id: i32, status: TaskStatus) -> TranscodeTask {
+fn make_task(id: i64, status: TaskStatus) -> TranscodeTask {
     TranscodeTask {
         id,
-        video_id: id as i64,
+        video_id: id,
         input_path: "/tmp/atmos-test-nonexistent.mp4".to_string(),
         resolutions: vec!["720p".to_string()],
         status,
